refactor: 예약 취소 승급 흐름 단순화
