+++ src/test/java/roomescape/reservation/ReservationServiceTest.java
@@ -16,7 +16,6 @@ import roomescape.domain.theme.Theme;
 import roomescape.exception.ConflictException;
 import roomescape.exception.InvalidInputException;
 import roomescape.exception.ResourceNotFoundException;
-import roomescape.service.reservation.ReservationCancellationService;
 import roomescape.service.reservation.ReservationService;
 import roomescape.service.reservationtime.ReservationTimeService;
 import roomescape.service.theme.ThemeService;
@@ -223,10 +222,6 @@ class ReservationServiceTest {
         private final FakeReservationSlotRepository reservationSlotRepository = new FakeReservationSlotRepository();
         private final FakeReservationWaitingRepository reservationWaitingRepository = new FakeReservationWaitingRepository();
         private final FakeReservationTimeRepository reservationTimeRepository = new FakeReservationTimeRepository();
-        private final ReservationCancellationService reservationCancellationService = new ReservationCancellationService(
-                reservationRepository,
-                reservationWaitingRepository
-        );
         private final ThemeService themeService = new ThemeService(themeRepository, reservationRepository);
         private final ReservationTimeService reservationTimeService = new ReservationTimeService(
                 reservationTimeRepository,
@@ -237,7 +232,7 @@ class ReservationServiceTest {
         private final ReservationService reservationService = new ReservationService(
                 reservationRepository,
                 reservationSlotRepository,
-                reservationCancellationService,
+                reservationWaitingRepository,
                 reservationTimeService,
                 themeService
         );
